+++ Soulmate/Soulmate/Domain/Interface/ChattingRepository.swift
@@ -5,14 +5,19 @@
 //  Created by Hoen on 2022/12/06.
 //
 
+import Combine
 import FirebaseFirestore
 
 protocol ChattingRepository {
+    var othersMessages: PassthroughSubject<[Chat], Never> { get }
+    
     var startDocument: QueryDocumentSnapshot? { get }
     var lastDocument: QueryDocumentSnapshot? { get }
     
     func setStartDocument(_ doc: QueryDocumentSnapshot?)
     func setLastDocument(_ doc: QueryDocumentSnapshot?)
+    
+    func removeListen()
     
     func loadReadChattings(from chatRoomId: String) async -> [MessageInfoDTO]
     func loadUnReadChattings(from chatRoomId: String) async -> [MessageInfoDTO]
@@ -21,6 +26,6 @@ protocol ChattingRepository {
     func updateUnreadCountToZero(of chatRoomId: String, othersId: String)
     func increaseUnreadCount(of id: String, in chatRoomId: String) async
     func addMessage(_ message: MessageToSendDTO, to chatRoomId: String) async -> Bool
-    func listenOthersChattingQuery(from chatRoomId: String) -> Query
+    func listenOthersChattings(from chatRoomId: String, uid: String)
     func listenOtherIsReading(from chatRoomId: String, userId: String) -> Query
 }
